feat: pack the schedule id, fees, admin, scalars, and dex percents (R50)
DcaSchedule drops to two slots: the keccak(user, token, nonce) id becomes the uint64
creation nonce itself, stored beside period, timestamp, route, and paused. Ids now start
at 1 and getSchedulesCreatedCount() is the last one assigned. Fee settings pack into two
slots (collector + both rates, then two uint128 bounds), OperationsAdmin holds handler and
deposit pause in one mapping value, DcaManager's protocol scalars and the nonce share one
slot, and the two Uniswap slippage percents share another. Every OperationsAdmin
route-index argument now carries the same toUint32() bound as registerRoute.

Create −30,459 gas, single purchase −6,207, five-row batch −78,588. Runtime grows
~1,450 bytes on DcaManager and the dex handlers, whose default-profile EIP-170 margins
fall to 1,991 and 1,544.

+++ src/interfaces/IFeeHandler.sol
@@ -10,11 +10,14 @@ interface IFeeHandler {
     ////////////////////////
     // Type declarations ///
     ////////////////////////
+    /// @dev Widths match FeeHandler storage: rates are capped at 5% (MAX_FEE_RATE_CAP) so they fit
+    ///      uint16, and the bounds are purchase amounts, so they carry the schedule's uint128.
+    ///      Setters and the constructor still take uint256 and SafeCast at the write.
     struct FeeSettings {
-        uint256 minFeeRate; // the lowest possible fee
-        uint256 maxFeeRate; // the highest possible fee
-        uint256 feePurchaseLowerBound; // the purchase amount below which max fee is applied
-        uint256 feePurchaseUpperBound; // the purchase amount above which min fee is applied
+        uint16 minFeeRate; // the lowest possible fee
+        uint16 maxFeeRate; // the highest possible fee
+        uint128 feePurchaseLowerBound; // the purchase amount below which max fee is applied
+        uint128 feePurchaseUpperBound; // the purchase amount above which min fee is applied
     }
 
     //////////////////////
